requirements.txt
streamlit
openai-whisper
pandas
numpy
soundfile
python-Levenshtein
torch
